+++ packages/web3/test/e2e/sepolia/get_code.test.ts
@@ -14,7 +14,7 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { bytesToBuffer, hexToBytes } from 'web3-utils';
+import { bytesToUint8Array, hexToBytes } from 'web3-utils';
 
 import Web3, { FMT_BYTES, FMT_NUMBER } from '../../../src';
 import {
@@ -65,7 +65,7 @@ describe(`${getSystemTestBackend()} tests - getCode`, () => {
 				break;
 			case 'BYTES_BUFFER':
 				// eslint-disable-next-line jest/no-conditional-expect
-				expect(result).toStrictEqual(bytesToBuffer(hexToBytes(expectedCode)));
+				expect(result).toStrictEqual(bytesToUint8Array(hexToBytes(expectedCode)));
 				break;
 			case 'BYTES_UINT8ARRAY':
 				// eslint-disable-next-line jest/no-conditional-expect